Only BuildConfiguration config needs skia
