+++ skia-bindings/build_support/binary_cache/binaries.rs
@@ -1,7 +1,7 @@
 //! Support for exporting and building prebuilt binaries.
 
 use super::{git, github_actions};
-use crate::build_support::{cargo, skia};
+use crate::build_support::{cargo, binaries_config};
 use flate2::read::GzDecoder;
 use std::{
     fs,
@@ -21,7 +21,7 @@ pub fn should_export() -> Option<PathBuf> {
 ///
 /// `source_files` are additional files from below skia-bindings/ that are copied to the target directory.
 pub fn export(
-    config: &skia::BinariesConfiguration,
+    config: &binaries_config::BinariesConfiguration,
     source_files: &[(&str, &str)],
     target_dir: &Path,
 ) -> io::Result<()> {
@@ -39,11 +39,7 @@ pub fn export(
 
     let target = cargo::target();
 
-    for lib in &config.ninja_built_libraries {
-        let filename = &target.library_to_filename(lib);
-        fs::copy(output_directory.join(filename), export_dir.join(filename))?;
-    }
-    for lib in &config.other_built_libraries {
+    for lib in config.built_libraries() {
         let filename = &target.library_to_filename(lib);
         fs::copy(output_directory.join(filename), export_dir.join(filename))?;
     }
