+++ skia-bindings/build_support/binary_cache/export.rs
@@ -1,7 +1,7 @@
 use std::path::Path;
 
 use crate::build_support::binary_cache::{binaries, SKIA_LICENSE, SRC_BINDINGS_RS};
-use crate::build_support::skia::BinariesConfiguration;
+use crate::build_support::binaries_config::BinariesConfiguration;
 
 /// Publish the binaries to Azure.
 pub fn publish(binaries_config: &BinariesConfiguration, staging_directory: &Path) {
