+++ skia-bindings/build_support/skia.rs
@@ -1,28 +1,19 @@
-//! Full build support for the Skia library, SkiaBindings library and bindings.rs file.
+//! Full build support for the Skia library.
 
-use crate::build_support::{android, cargo, clang, features, ios, llvm, vs};
+use crate::build_support::{android, binaries_config, cargo, clang, features, ios, llvm, vs};
 use std::env;
 use std::path::{Path, PathBuf};
 use std::process::{Command, Stdio};
 
-/// The libraries to link with.
-mod lib {
-    pub const SKIA: &str = "skia";
-    pub const SK_SHAPER: &str = "skshaper";
-    pub const SK_PARAGRAPH: &str = "skparagraph";
-}
-
-/// The defaults for the Skia build configuration.
-impl From<features::Features> for BuildConfiguration {
-    fn from(features: features::Features) -> Self {
-        let skia_debug = matches!(cargo::env_var("SKIA_DEBUG"), Some(v) if v != "0");
-
+/// Builds a Skia configuration from a Features set.
+impl BuildConfiguration {
+    pub fn from_features(features: features::Features, skia_debug: bool) -> Self {
         BuildConfiguration {
             on_windows: cargo::host().is_windows(),
-            skia_debug,
             // `OPT_LEVEL` is set by Cargo itself.
             opt_level: cargo::env_var("OPT_LEVEL"),
             features,
+            skia_debug,
             cc: cargo::env_var("CC").unwrap_or_else(|| "clang".to_string()),
             cxx: cargo::env_var("CXX").unwrap_or_else(|| "clang++".to_string()),
         }
@@ -301,146 +292,10 @@ fn yes_if(y: bool) -> String {
     }
 }
 
-/// The configuration of the resulting binaries.
-#[derive(Clone, PartialEq, Eq, Debug)]
-pub struct BinariesConfiguration {
-    /// The feature identifiers we built with.
-    pub feature_ids: Vec<String>,
-
-    /// The output directory of the libraries we build and we need to inform cargo about.
-    pub output_directory: PathBuf,
-
-    /// The TARGET specific link libraries we need to inform cargo about.
-    pub link_libraries: Vec<String>,
-
-    /// The static Skia libraries skia-bindings provides and dependent projects need to link with.
-    pub ninja_built_libraries: Vec<String>,
-
-    /// The static Skia libraries skia-bindings provides and dependent projects need to link with.
-    pub other_built_libraries: Vec<String>,
-
-    /// Additional files relative to the output_directory
-    /// that are needed to build dependent projects.
-    pub additional_files: Vec<PathBuf>,
-
-    /// `true` if the skia libraries are built with debugging information.
-    pub skia_debug: bool,
-}
-
-const SKIA_OUTPUT_DIR: &str = "skia";
-const ICUDTL_DAT: &str = "icudtl.dat";
-
-impl BinariesConfiguration {
-    /// Build a binaries configuration based on the current environment cargo
-    /// supplies us with and a Skia build configuration.
-    pub fn from_cargo_env(build: &BuildConfiguration) -> Self {
-        let features = &build.features;
-        let target = cargo::target();
-
-        let mut ninja_built_libraries = Vec::new();
-        let mut additional_files = Vec::new();
-        let feature_ids = features.ids();
-
-        if features.text_layout {
-            if target.is_windows() {
-                additional_files.push(ICUDTL_DAT.into());
-            }
-            ninja_built_libraries.push(lib::SK_PARAGRAPH.into());
-            ninja_built_libraries.push(lib::SK_SHAPER.into());
-        }
-
-        let mut link_libraries = Vec::new();
-
-        match target.as_strs() {
-            (_, "unknown", "linux", _) => {
-                link_libraries.extend(vec!["stdc++", "fontconfig", "freetype"]);
-                if features.gl {
-                    if features.egl {
-                        link_libraries.push("EGL");
-                    }
-
-                    if features.x11 {
-                        link_libraries.push("GL");
-                    }
-
-                    if features.wayland {
-                        link_libraries.push("wayland-egl");
-                        link_libraries.push("GLESv2");
-                    }
-                }
-            }
-            (_, "apple", "darwin", _) => {
-                link_libraries.extend(vec!["c++", "framework=ApplicationServices"]);
-                if features.gl {
-                    link_libraries.push("framework=OpenGL");
-                }
-                if features.metal {
-                    link_libraries.push("framework=Metal");
-                    // MetalKit was added in m87 BUILD.gn.
-                    link_libraries.push("framework=MetalKit");
-                    link_libraries.push("framework=Foundation");
-                }
-            }
-            (_, _, "windows", Some("msvc")) => {
-                link_libraries.extend(&["usp10", "ole32", "user32", "gdi32", "fontsub"]);
-                if features.gl {
-                    link_libraries.push("opengl32");
-                }
-                if features.d3d {
-                    link_libraries.extend(&["d3d12", "dxgi", "d3dcompiler"]);
-                }
-            }
-            (_, "linux", "android", _) | (_, "linux", "androideabi", _) => {
-                link_libraries.extend(android::link_libraries(features));
-            }
-            (_, "apple", "ios", _) => {
-                link_libraries.extend(ios::link_libraries(features));
-            }
-            _ => panic!("unsupported target: {:?}", cargo::target()),
-        };
-
-        let output_directory = cargo::output_directory()
-            .join(SKIA_OUTPUT_DIR)
-            .to_str()
-            .unwrap()
-            .into();
-
-        ninja_built_libraries.push(lib::SKIA.into());
-
-        BinariesConfiguration {
-            feature_ids: feature_ids.into_iter().map(|f| f.to_string()).collect(),
-            output_directory,
-            link_libraries: link_libraries
-                .into_iter()
-                .map(|lib| lib.to_string())
-                .collect(),
-            ninja_built_libraries,
-            other_built_libraries: vec![],
-            additional_files,
-            skia_debug: build.skia_debug,
-        }
-    }
-
-    /// Inform cargo that the library files of the given configuration are available and
-    /// can be used as dependencies.
-    pub fn commit_to_cargo(&self) {
-        cargo::add_link_search(self.output_directory.to_str().unwrap());
-
-        // On Linux, the order is significant, first the static libraries we built, and then
-        // the system libraries.
-
-        let target = cargo::target();
-
-        cargo::add_static_link_libs(&target, &self.ninja_built_libraries);
-        cargo::add_static_link_libs(&target, &self.other_built_libraries);
-        cargo::add_link_libs(&self.link_libraries);
-    }
-}
-
 /// Orchestrates the entire build of Skia based on the arguments provided.
 pub fn build(
     build: &FinalBuildConfiguration,
-    config: &BinariesConfiguration,
+    config: &binaries_config::BinariesConfiguration,
     ninja_command: Option<PathBuf>,
     gn_command: Option<PathBuf>,
     offline: bool,
@@ -478,7 +333,7 @@ pub fn build(
 /// Configures Skia by calling gn
 pub fn configure_skia(
     build: &FinalBuildConfiguration,
-    config: &BinariesConfiguration,
+    config: &binaries_config::BinariesConfiguration,
     python2: &Path,
     gn_command: Option<&Path>,
 ) {
@@ -518,7 +373,7 @@ pub fn configure_skia(
 ///
 /// This function assumes that all prerequisites are in place and that the output directory
 /// contains a fully configured Skia source tree generated by gn.
-pub fn build_skia(config: &BinariesConfiguration, ninja_command: &Path) {
+pub fn build_skia(config: &binaries_config::BinariesConfiguration, ninja_command: &Path) {
     let ninja_status = Command::new(ninja_command)
         .args(&config.ninja_built_libraries)
         .args(&["-C", config.output_directory.to_str().unwrap()])
